requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
bootstrap4==0.1.0
colorama==0.4.3
Django>=3.0.2
django-braces==1.14.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-summernote==0.8.11.6
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
pycodestyle==2.5.0
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
virtualenv==16.7.4
wrapt==1.11.2
